Move GradleRuntimeCompatibility implementation to internal package
The original public API is still available but deprecated.
