+++ subprojects/gradle-plugin-development/build.gradle
@@ -1,5 +1,3 @@
-import static dev.gradleplugins.GradleRuntimeCompatibility.*
-
 plugins {
     id 'org.jetbrains.kotlin.jvm' version "1.3.72"
     id 'toolboxbuild.publish'
